Move the sans-I/O stream types out of the LZMA2 module
Move the stream-related types to their own module so that they could
be used properly from XzStream, Lzma2Stream and LzmaStream
