+++ src/lib.rs
@@ -63,6 +63,7 @@ mod lzma2_reader;
 mod lzma_reader;
 mod range_dec;
 mod state;
+mod stream;
 #[cfg(feature = "std")]
 mod work_queue;
 #[cfg(feature = "xz")]
@@ -104,10 +105,7 @@ pub use lzma_reader::{
     LzmaReader, LzmaStream, get_memory_usage as lzma_get_memory_usage,
     get_memory_usage_by_props as lzma_get_memory_usage_by_props,
 };
-pub use lzma2_reader::{
-    Action, Lzma2Reader, Lzma2Stream, Status, StreamResult,
-    get_memory_usage as lzma2_get_memory_usage,
-};
+pub use lzma2_reader::{Lzma2Reader, Lzma2Stream, get_memory_usage as lzma2_get_memory_usage};
 #[cfg(feature = "std")]
 pub use lzma2_reader_mt::Lzma2ReaderMt;
 #[cfg(not(feature = "std"))]
@@ -117,6 +115,7 @@ pub use no_std::Read;
 #[cfg(not(feature = "std"))]
 pub use no_std::Write;
 use state::*;
+pub use stream::{Action, Status, StreamResult};
 #[cfg(all(feature = "xz", feature = "std"))]
 pub use xz::XzReaderMt;
 #[cfg(all(feature = "xz", feature = "encoder", feature = "std"))]
